requirements.txt
Flask==2.2.2
Flask-Session==0.4.0
Werkzeug==2.2.2
Flask_Redis==0.4.0
redis==4.3.4
pika==1.3.1
mysql-connector-python==8.0.30
PyJWT==2.3.0